requirements.txt
rumps==0.4.0
pystray==0.19.4
Pillow==10.0.0
pyyaml==6.0.1
pyperclip